requirements.txt
Django==2.0.5
djangorestframework==3.8.2
python-dateutil==2.7.3
coreschema==0.0.4
coreapi==2.3.3
django-filter==1.1.0
psycopg2-binary==2.7.4
django-environ==0.4.4
gunicorn==19.8.1
whitenoise==3.3.1
